Make the model overfit on the training set.
Apparently, there is a problem with training the model with batchnorm when the input data to the model is smaller than the dataloader minibatch_size, (i.e. for the last iteration of each epoch for the remaining data samples of the dataset which are smaller than a full minibatch.
For now I just disregarded this problem by not training on that partial mini-batch and moving on to the next epoch.